filter - keep statistics about delivered and dropped mail
Update the tests accordingly.  This requires a schema change, and the
filter needed to be refactored to allow the new fields to be left-as
is whilst updating the remaining count.  What I wanted was an upsert;
best answers here suggest a combination of insert and update queries
would be simpler.

https://stackoverflow.com/questions/418898/sqlite-upsert-not-insert-or-replace

Still not easy!