filter, data-acl, install - emulate $verify_mode with an acl variable
Set the variable in data acl, then only delivery will have it set.
With the committed changes to the filter, this allows us to avoid
decrementing the remaining counter twice, once in verify mode and once
in deliviery mode.

Still a bit of a hack.  If there is another custom acl defined using
CHECK_DATA_LOCAL_ACL_FILE then this will clash.