package - don't use `service` in maintainer scripts
I need to learn more about the dh_make process, probably, but for now
I realise using `service` is wrong, and `reload` is also wrong, as it
fails when the service is not running.

The init script /etc/init.d/exim4 is distributed with exim and has
been forever, so use that.  Systemd can probably be ignored for now,
as this script should handled it accordingly.